pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = []

[project]
name = "binocularslive"
description = "API Portal for Binoculars"
readme = "README.md"
authors = [ {name="Marc Schwarzschild", email="ms@TheBrookhavenGroup.com"}, ]
maintainers = [ {name="Marc Schwarzschild", email="ms@TheBrookhavenGroup.com"}, ]
requires-python = ">=3.12"
keywords = ["binoculars, llm"]
classifiers = []
license = {file="LICENSE"}
dependencies = [
    "django",
    "psycopg2-binary",
    "tbgutils",
    "pycodestyle",
    "djangorestframework",
]
version = "0.0.1"

[project.urls]
homepage = "https://github.com/thebrookhavengroup/binocularslive"
repository = "https://github.com/thebrookhavengroup/binocularslive"
documentation = "https://github.com/thebrookhavengroup/binocularslive"

[project.scripts]
binoculars = "binocularslive.manage:main"


